requirements.txt
langgraph==0.0.10
langchain==0.1.0
langchain-community==0.0.10
pinecone-client==3.0.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
pytest==7.4.4
termcolor==3.1.0
pydantic==2.12.0
google-genai>=1.40.0
playwright==1.55.0
browserbase==1.4.0
rich